Understanding the Problem Statement
After talking to the Apple Engineer at the Lab implemented an Initialiser for an Optional Array but got errors regarding StoryBoard.
If you are using a Storyboard for your views then a decoder is necessary to initialise the view.In that case initialising an empty variable is not Possible.
I also got to learn to use Compositional Layout and will be implementing it's use cases really soon